Upgrades, Enemy Types, and Waves
Towers can now be selected by being clicked on, selecting shows range and allows upgrading fire rate, range, and damage by pressing 1, 2, and 3, respectively. Upgrading one of a tower's stats doubles the stat and changes the tower's color accordingly. Pressing the right mouse button activates build mode, and causes an overlay to follow the mouse and show the range of a tower if it were at the mouse position. Towers can now only be built in build mode. Two new enemy types have been added: Darts, which are orange, move faster and have less health. Brutes, which are purplish, move slower and have more health. Darts can be test spawned wiwth w, brutes with q. Pressing spacebar now starts a wave, which spawns a preset number of enemies with set types and spawn timings. The wave ends when all enemies in that wave have spawed and either have reached the end or been destroyed, upon which the player recieves 10 money and becomes able to spawn the next wave. A tracker at the top of the screen shows what the current wave is.